:bug: Change quantity type to decimal

+++ data/migrations/20191107141731_create_tables.js
@@ -27,7 +27,7 @@ exports.up = function(knex) {
         .inTable("ingredients");
 
       table
-        .integer("quantity")
+        .decimal("quantity")
         .unsigned()
         .notNullable();
     })
